+++ apps/mb-api/src/objectifAvancement/commands/upsertObjectifAvancement.ts
@@ -4,14 +4,10 @@ import {
 } from '@pilote/mb-shared/objectifAvancement'
 import { ResultAsync } from 'neverthrow'
 
-import { requireCurrentPrincipalId } from '@/framework/auth/userContext'
 import { Decimal } from '@/framework/decimal'
 import { db } from '@/framework/persistence/dbStore'
-import { type IndividuInconnuError, resolveAuthorizedIndividu } from '@/individu/permission'
-import {
-  ensureIndicateurWritePermission,
-  withIndicateurReadPermission,
-} from '@/indicateur/permissions'
+import { type IndividuInconnuError } from '@/individu/permission'
+import { resolveIndicateurAndIndividu } from '@/indicateur/resolveIndicateurAndIndividu'
 import { toObjectifAvancementApiModel } from '@/objectifAvancement/utils'
 
 export type UpsertObjectifAvancementError = IndividuInconnuError
@@ -27,48 +23,31 @@ export const upsertObjectifAvancement = ({
 }: UpsertObjectifAvancementParams): ResultAsync<
   ObjectifAvancementApiModel,
   UpsertObjectifAvancementError
-> => {
-  const principalId = requireCurrentPrincipalId()
-  return ResultAsync.fromSafePromise(
-    db().indicateur.findFirstOrThrow({
-      where: withIndicateurReadPermission({ publicId: indicateurPublicId }, principalId),
-      select: { id: true, publicId: true },
-    }),
-  )
-    .andThen((indicateur) =>
-      ensureIndicateurWritePermission({ indicateurId: indicateur.id, principalId }).map(
-        () => indicateur,
-      ),
-    )
-    .andThen((indicateur) =>
-      resolveAuthorizedIndividu({
-        individuPublicId: body.individu,
-        indicateurId: indicateur.id,
-      }).map((individu) => ({ indicateur, individu })),
-    )
-    .andThen(({ indicateur, individu }) =>
-      ResultAsync.fromSafePromise(
-        db().objectifAvancement.upsert({
-          where: {
-            objectif_avancement_unique: {
-              indicateurId: indicateur.id,
-              individuId: individu.id,
-              date: body.date,
-            },
-          },
-          update: { valeur: new Decimal(body.valeur) },
-          create: {
+> =>
+  resolveIndicateurAndIndividu({
+    indicateurPublicId,
+    individuPublicId: body.individu,
+  }).andThen(({ indicateur, individu }) =>
+    ResultAsync.fromSafePromise(
+      db().objectifAvancement.upsert({
+        where: {
+          objectif_avancement_unique: {
             indicateurId: indicateur.id,
             individuId: individu.id,
             date: body.date,
-            valeur: new Decimal(body.valeur),
           },
-          include: {
-            indicateur: { select: { publicId: true } },
-            individu: { select: { publicId: true } },
-          },
-        }),
-      ),
-    )
-    .map(toObjectifAvancementApiModel)
-}
+        },
+        update: { valeur: new Decimal(body.valeur) },
+        create: {
+          indicateurId: indicateur.id,
+          individuId: individu.id,
+          date: body.date,
+          valeur: new Decimal(body.valeur),
+        },
+        include: {
+          indicateur: { select: { publicId: true } },
+          individu: { select: { publicId: true } },
+        },
+      }),
+    ).map(toObjectifAvancementApiModel),
+  )
