+++ apps/mb-api/src/valeurAvancement/commands/deleteValeurAvancement.ts
@@ -1,13 +1,9 @@
 import { type DeleteValeurAvancementBody } from '@pilote/mb-shared/valeurAvancement'
 import { ResultAsync } from 'neverthrow'
 
-import { requireCurrentPrincipalId } from '@/framework/auth/userContext'
 import { db } from '@/framework/persistence/dbStore'
-import { type IndividuInconnuError, resolveAuthorizedIndividu } from '@/individu/permission'
-import {
-  ensureIndicateurWritePermission,
-  withIndicateurReadPermission,
-} from '@/indicateur/permissions'
+import { type IndividuInconnuError } from '@/individu/permission'
+import { resolveIndicateurAndIndividu } from '@/indicateur/resolveIndicateurAndIndividu'
 
 export type DeleteValeurAvancementError = IndividuInconnuError
 
@@ -19,35 +15,18 @@ type DeleteValeurAvancementParams = {
 export const deleteValeurAvancement = ({
   indicateurPublicId,
   body,
-}: DeleteValeurAvancementParams): ResultAsync<void, DeleteValeurAvancementError> => {
-  const principalId = requireCurrentPrincipalId()
-  return ResultAsync.fromSafePromise(
-    db().indicateur.findFirstOrThrow({
-      where: withIndicateurReadPermission({ publicId: indicateurPublicId }, principalId),
-      select: { id: true, publicId: true },
-    }),
+}: DeleteValeurAvancementParams): ResultAsync<void, DeleteValeurAvancementError> =>
+  resolveIndicateurAndIndividu({
+    indicateurPublicId,
+    individuPublicId: body.individu,
+  }).andThen(({ indicateur, individu }) =>
+    ResultAsync.fromSafePromise(
+      db().valeurAvancement.deleteMany({
+        where: {
+          indicateurId: indicateur.id,
+          individuId: individu.id,
+          date: body.date,
+        },
+      }),
+    ).map(() => undefined),
   )
-    .andThen((indicateur) =>
-      ensureIndicateurWritePermission({ indicateurId: indicateur.id, principalId }).map(
-        () => indicateur,
-      ),
-    )
-    .andThen((indicateur) =>
-      resolveAuthorizedIndividu({
-        individuPublicId: body.individu,
-        indicateurId: indicateur.id,
-      }).map((individu) => ({ indicateur, individu })),
-    )
-    .andThen(({ indicateur, individu }) =>
-      ResultAsync.fromSafePromise(
-        db().valeurAvancement.deleteMany({
-          where: {
-            indicateurId: indicateur.id,
-            individuId: individu.id,
-            date: body.date,
-          },
-        }),
-      ),
-    )
-    .map(() => undefined)
-}
